resolve duplicate files
