+++ src/Components/styled/activity-detail-comp/ActivityDetail.styled.jsx
