+++ src/Components/styled/activity-detail-comp/Middle.styled.jsx
