--- src/components/Footer.jsx
@@ -1,49 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-import BondiLogo from "../assets/icons/second-logo.svg?react";
-import {
-  FooterInner,
-  FooterWrapper,
-  LeftBlock,
-  Tagline,
-  CenterBlock,
-  RightBlock,
-  FooterLink,
-  LogoContainer,
-} from "./styled/Footer.styled.jsx";
-
-function Footer() {
-  return (
-    <FooterWrapper>
-      <FooterInner>
-        <LeftBlock>
-          <LogoContainer>
-            <BondiLogo />
-          </LogoContainer>
-          <Tagline>
-            {" "}
-            Connect through shared activities and build meaningful friendships
-            in your local community.
-          </Tagline>
-        </LeftBlock>
-
-        <CenterBlock>
-          <FooterLink as={Link} to="/contact">
-            {" "}
-            Contact Us{" "}
-          </FooterLink>
-          <FooterLink as={Link} to="/about">
-            {" "}
-            About Us{" "}
-          </FooterLink>
-        </CenterBlock>
-
-        <RightBlock>
-          &copy; {new Date().getFullYear()} Bondi. All rights reserved.
-        </RightBlock>
-      </FooterInner>
-    </FooterWrapper>
-  );
-}
-
-export default Footer;
