+++ src/pages/Profile.jsx
@@ -1,15 +1,15 @@
-import Footer from "../Components/Footer";
-import NavBar from "../Components/NavBar";
-import Cover from "../Components/profile_components/Cover";
-import About from "../Components/profile_components/About";
-import Interest from "../Components/profile_components/Interest";
-import RecentActivities from "../Components/profile_components/RecentActivities";
-import ActivityStatus from "../Components/profile_components/ActivityStatus";
+import Footer from "../components/feed-components/Footer";
+import NavBar from "../components/feed-components/NavBar";
+import Cover from "../components/profile_components/Cover";
+import About from "../components/profile_components/About";
+import Interest from "../components/profile_components/Interest";
+import RecentActivities from "../components/profile_components/RecentActivities";
+import ActivityStatus from "../components/profile_components/ActivityStatus";
 import {
   ProfilePageContainer,
   ProfileMainContent,
-  ProfileSectionsContainer
-} from "../Components/styled/profile-style-comp/Profile.styled";
+  ProfileSectionsContainer,
+} from "../components/styled/profile-style-comp/Profile.styled";
 
 export default function Profile() {
   return (
